+++ src/textual_fastdatatable/__main__.py
@@ -3,10 +3,12 @@ from textual.app import App, ComposeResult
 from textual_fastdatatable import ArrowBackend, DataTable
 
 
-class TableApp(App):
+class TableApp(App, inherit_bindings=False):
+    BINDINGS = [("ctrl+q", "quit", "Quit")]
+
     def compose(self) -> ComposeResult:
         backend = ArrowBackend.from_parquet("./tests/data/lap_times_538121.parquet")
-        yield DataTable(backend=backend)
+        yield DataTable(backend=backend, cursor_type="range")
 
 
 if __name__ == "__main__":